runtime: report inlined frames in Frames.Next
This CL is a sketch of a change to the stack frame iterator
that makes it report inlined frames too. Before, we observed
that the sequence of Frame.Function strings would omit
inlined frames that should be present in the logical call stack.

For example, issue #65516 shows a telemetry-reported call
stack that includes these frames:

	...
	go/ast.Walk
	go/ast.Inspect
	golang.org/x/tools/gopls/internal/golang.highlightPath
	...

However, there is no direct call from highlightPath to Inspect:
the intervening highlightFuncControlFlow frame has been elided.
(The call from highlightPath to highlightFuncControlFlow was
not inlined, by the call from highlightFuncControlFlow to
Inspect was.)

The sequence of PC values provided to CallersFrames was
obtained not from runtime.Callers, but was extracted from
the pc=0x%x strings appearing in the GOTRACEBACK=system
output, which include these:

	panic = 10262acb0
	panicmem: no pc (inlined)
 	runtime.sigpanic = 1026457d8
	go/ast.Walk = 1027bfa20
	go/ast.Inspect: no pc (inlined)
	golang.org/x/tools/gopls/internal/golang.highlightFuncControlFlow = 102cd82e8
	golang.org/x/tools/gopls/internal/golang.highlightPath = 102cd7c98
	...

This change causes the stack to report the complete set of
logical frames, and fixes my problem. That said,
I suspect that there is no actual bug in combination of Callers
+ CallersFrames because Callers already applied the inline
unwinder, so CallersFrames does not need to. But if that's the
case then the GOTRACEBACK=system output seems to be logically
incomplete, because it doesn't render the pcs for the inlined
frames. Perhaps the appropriate fix is there instead.

WORK IN PROGRESS
DO NOT SUBMIT

Change-Id: Ia1b7180e154a33d8942a9469401db4289e02d0e8