cmd/compile: caching result of AlgType, EqCanPanic and PtrDataSize
Preventing exponential behavior when compiling deeply nested structs.
For some program with 33 nested types, the compile time down from 10m to 10ms.

This CL is based on CL 561875.

I'm not sure if caching the smartest thing to do here, maybe theses values
could be moved inside the type struct and computed when the type is created
but whatever this works and does the job.

Fixes #65540

Change-Id: Idced67b57d96ae27f353075a8d75ef1ab1b41d01