runtime: don't call lockOSThread for every syscall call on Windows
When making system calls, Windows passes arguments in a struct held in
the m to save some stack space, which is very restricted on nosplit
functions. This requires a costly call to lockOSThread to ensure that
the m remains the same during the syscall.

This CL changes the Windows syscall implementation to allocate
the syscall arguments on the stack, adding 48 bytes of stack per call.
It then partially compensates this stack growths by not requiring
to have at least 32 bytes of stack when calling asmstdcall, so
the final overhead is just 16 bytes per syscall.

With this we can remove the lockOSThread call, which is a source
of context switches. It also give a small performance boost to
syscalls:

goos: windows
goarch: amd64
pkg: runtime
cpu: Intel(R) Core(TM) i7-10850H CPU @ 2.70GHz
                 │   old.txt   │              new.txt               │
                 │   sec/op    │   sec/op     vs base               │
SyscallSimple-12   70.92n ± 2%   64.92n ± 3%  -8.47% (p=0.000 n=40)

Updates #21827.
Updates #58336.

Change-Id: If6644fd111dbacab74e7dcee2afa18ca146735da