runtime: merge timerRemoving into timerModifying
timerRemoving is just a kind of "locked for modification",
so merge it into timerModifying. This does potentially remove
a fast path from deltimer, in that deltimer of timerRemoving
is a fast-path exit while deltimer of timerModifying has to
wait for the timer to settle. Since all the timerModifying
critical paths are bounded and short, this should not matter.

This is part of a larger simplification of the state set.

[This is one CL in a refactoring stack making very small changes
in each step, so that any subtle bugs that we miss can be more
easily pinpointed to a small change.]

OK TO REVIEW

DO NOT SUBMIT: Double-check that the deltimer issue is not
a potential source of deadlocks. There is a staticlockranking
issue identified in a later CL, which is almost certainly a larger
issue than just this one case, but wait to submit until we understand
that better.

Change-Id: I039bf6a5a041a158dc3d1af8127f28eed50fc540