runtime: use timer.lock in deltimer
The state set is now simplified enough that all the CAS loops
are starting to look the same: they are just spin locks.
So introduce an actual timer.lock method and use it in deltimer.

[This is one CL in a refactoring stack making very small changes
in each step, so that any subtle bugs that we miss can be more
easily pinpointed to a small change.]

OK TO REVIEW

DO NOT SUBMIT: Move staticlockranking CL down into this one,
and then resolve the potential deadlocks, either in this CL or a followup.
(The deadlocks do not appear to be new in this refactoring,
but figure out the plan for them first.)

Change-Id: Ifd7f20eeede5c764ef10ecba64855c29a5ddbe39