cmd/asm,cmd/compile: generate less instructions for most 32 bit constant adds on ppc64x
For GOPPC64 < 10 targets, most large 32 bit constants (those
exceeding int16 capacity) can be added in two instructions
instead of generating a constant in REGTMP and adding it (which
takes 3 instructions).

It does not work for 32 bit values greater than 0x7FFF7FFF,
so this must be done during asm preprocessing as the optab
matching rules cannot differentiate this special case.

Note, constants 0x10000 > x >= 0x8000 are not converted. The
ORI + ADD sequence is faster than ADDIS + ADD for the same
number of instructions.

Change-Id: I1ccc839c6c28fc32f15d286b2e52e2d22a2a06d4