internal/trace/v2: clean up the ordering.advance interface
This change cleans up ordering.advance in three major ways.

First, it resolves a TODO about using a proper queue of events for extra
events produced while performing ordering. This will be necessary for a
follow-up change to handle coroutine switch events.

Next, it simplifies the ordering.advance method's signature by returning
an Event instead of a schedCtx and a bool indicating whether or not to
advance. This means that ordering.advance is now responsible for
constructing the event instead of the caller, which makes the caller
side a bit easier to understand.

Finally, hand-in-hand with the signature change, the implementation of
ordering.advance no longer forces each switch case to return but instead
has them converge past the switch. This has two effects. One is that we
eliminate the deferred call to update the M state. Using a defer here is
technically incorrect, because we might end up changing the M state even
if we don't advance the event! We got lucky here that curCtx == newCtx
in all such cases, but there may have been a subtle bug lurking here. A
secondary effect of this is that we have a place to construct the Event
value that ordering.advance returns, so not every advancing return path
needs to construct the Event.

Change-Id: Ifebbbf14e8ed5c08be5c1b0fadc2e5df3915c656