go/types, types2: introduce Pos alias for syntax.Pos/token.Pos
DO NOT SUBMIT.
Experiment.

This introduces a new exported name: go/types/Pos. We can either
accept that (with proposal) or work around it by using _Pos or
aPos or the like.

Using a Pos alias shortens a lot of code and removes the need
for many rewrites for the auto-generated go/types files.

Also, declare isKnown function to be used instead of the isKnown
or IsValid Pos methods.

Change-Id: Ifac6cbf5456f997f5f73d099ea846e36889d7682