runtime: emit trace stacks for more goroutines in each generation
This change adds a new event, GoStatusStack, which is like GoStatus but
also carries a stack ID. The purpose of this event is to emit stacks in
more places, in particular for goroutines that may never emit a
stack-bearing event in a whole generation.

There are two broad categories of cases handled by this CL.

The first is goroutines that were blocked or in a syscall the entire
generation. This particular case is handled at the point that we scribble
down the goroutine's status before the generation transition. That way,
when we're finishing up the generation and emitting events for any
goroutines we scribbled down, we have an accurate stack for those
goroutines ready to go, and we emit a GoStatusStack instead of a
GoStatus event. There's a small drawback with the way we scribble down
the stack though: we immediately register it in the stack table instead
of tracking the PCs. This means that if a goroutine does run and emit a
trace event in between when we scribbled down its stack and the end of
the generation, we will have recorded a stack that never actually gets
referenced in the trace. This case should be rare.

The second is goroutines that were unblocked but either never got a
chance to run within the generation, or just never emitted an event once
they started running. In this case, we emit a GoStatusStack event
instead of a GoStatus event as well, for the about-to-be-unblocked
goroutine.

Note that many other cases are covered by the fact that running
goroutines are preempted once per trace generation.

The only remaining case is a goroutine that was runnable when tracing
started and began running. Then, it ran til the end of the generation
without getting preempted. This case is rare, but still certainly
possible. I believe the only way to resolve it is to emit a
GoStatusStack event instead of a GoStatus event every time. We may
consider that in the future, but for now, this substantially improves
the current situation.

Fixes #65634.

Change-Id: Ief3b6df5848b426e7ee6794e98dc7ef5f37ab2d0