runtime: restructure scanobject loop to be more CPU friendly
This change restructures the scanobject loop to buffer pointers from
multiple small objects before processing them all together for marking.
The goal behind this change is to hide the latency of each (likely)
cache miss. It also makes other small changes like structuring loops to
give the CPU some idea of when the loop will likely end, reducing branch
mispredictions.

The downside of this change is that it does add overhead to the core GC
scan loop, so if the entire heap fits in, for example, L2 or L3 cache,
then this change is less useful. However, that is more likely to only be
a problem for microbenchmarks.

Change-Id: I59690da26f218e45d90c29e3ed52a5c0ef425b10