cmd/compile: fix sign/zero-extension removal
When an opcode generates a known high bit state (typically, a
sub-word operation that zeros the high bits), we can remove any
subsequent extension operation that would be a no-op.

x = (OP ...)
y = (ZeroExt32to64 x)

If OP zeros the high 32 bits, then we can replace y with x, as the
zero extension doesn't do anything.

However, x in this situation normally has a sub-word-sized type.  The
semantics of values in registers is typically that the high bits
beyond the value's type size are junk. So although the opcode
generating x *currently* zeros the high bits, after a rewrite it may
not - rewrites of sub-word-sized values can trash the high bits.

This is also a worry if x is spilled and restored. Spilling and
restoring doesn't preserve the high bits, but instead sets them to a
known value (often 0, but in some cases it could be sign-extended).

Instead, when we remove the extension op, set the type of the argument
value to a full-width type. That ensures that the op will correctly
generate the full-width result through any subsequent optimizations,
and any spill/restore path.

Fixes #66066

Change-Id: I3b5c091b3b3278ccbb7f11beda8b56f4b6d3fde7